+++ Sources/PerfectHTTPServer/HTTP2Request.swift
@@ -1,0 +1,23 @@
+//
+//  HTTP2.swift
+//  PerfectLib
+//
+//  Created by Kyle Jessup on 2016-02-18.
+//  Copyright © 2016 PerfectlySoft. All rights reserved.
+//
+//===----------------------------------------------------------------------===//
+//
+// This source file is part of the Perfect.org open source project
+//
+// Copyright (c) 2015 - 2016 PerfectlySoft Inc. and the Perfect project authors
+// Licensed under Apache License v2.0
+//
+// See http://perfect.org/licensing.html for license information
+//
+//===----------------------------------------------------------------------===//
+//
+
+final class HTTP2Request: HTTP11Request {
+	
+	
+}
